attempt to test mqtt-connect, move refresh data-mappings
