+++ wis2box-management/wis2box/api/__init__.py
@@ -28,23 +28,9 @@ from time import sleep
 from wis2box import cli_helpers
 from wis2box.api.backend import load_backend
 from wis2box.api.config import load_config
-from wis2box.env import (DOCKER_BROKER, DOCKER_API_URL, API_URL)
-from wis2box.plugin import load_plugin, PLUGINS
+from wis2box.env import (DOCKER_API_URL, API_URL)
 
 LOGGER = logging.getLogger(__name__)
-
-
-def refresh_data_mappings():
-    # load plugin for local broker
-    defs_local = {
-        'codepath': PLUGINS['pubsub']['mqtt']['plugin'],
-        'url': DOCKER_BROKER,
-        'client_type': 'dataset-manager'
-    }
-    local_broker = load_plugin('pubsub', defs_local)
-    success = local_broker.pub('wis2box/data_mappings/refresh', '{}', qos=0)
-    if not success:
-        LOGGER.error('Failed to refresh data mappings')
 
 
 def execute_api_process(process_name: str, payload: dict) -> dict:
@@ -134,9 +120,6 @@ def setup_collection(meta: dict = {}) -> bool:
             LOGGER.error(msg)
             return False
 
-    LOGGER.info('Refreshing data mappings')
-    refresh_data_mappings()
-
     return True
 
 
@@ -182,9 +165,6 @@ def remove_collection(collection_id: str, backend: bool = True,
         except Exception:
             msg = f'discovery metadata {collection_id} not found'
             LOGGER.warning(msg)
-
-    LOGGER.info('Refreshing data mappings')
-    refresh_data_mappings()
 
     return True
