+++ wis2box-management/wis2box/pubsub/mqtt.py
@@ -21,6 +21,8 @@
 
 import logging
 import random
+import time
+
 from typing import Any, Callable
 
 from paho.mqtt import client as mqtt_client
@@ -52,8 +54,6 @@ class MQTTPubSubClient(BasePubSubClient):
 
         self.conn.enable_logger(logger=LOGGER)
 
-        self.is_connected = False
-
         if None not in [self.broker_url.password, self.broker_url.password]:
             self.conn.username_pw_set(
                 self.broker_url.username,
@@ -67,30 +67,8 @@ class MQTTPubSubClient(BasePubSubClient):
         if self.broker_url.scheme == 'mqtts':
             self.conn.tls_set(tls_version=2)
 
-        def on_connect(client, userdata, flags, rc):
-            # check if the connection was successful
-            if rc == 0:
-                LOGGER.debug(f'Connected to broker {self.broker}')
-                self.is_connected = True
-            else:
-                self.is_connected = False
-                msg = f"Connection to MQTT-broker with url={self.broker['url']} failed:"  # noqa	
-                if rc == mqtt_client.CONNACK_REFUSED_PROTOCOL_VERSION:
-                    reason = 'Connection Refused, unacceptable protocol version' # noqa
-                elif rc == mqtt_client.CONNACK_REFUSED_IDENTIFIER_REJECTED:
-                    reason = 'Connection Refused, identifier rejected'
-                elif rc == mqtt_client.CONNACK_REFUSED_SERVER_UNAVAILABLE:
-                    reason = 'Connection Refused, server unavailable'
-                elif rc == mqtt_client.CONNACK_REFUSED_BAD_USERNAME_PASSWORD:
-                    reason = 'Connection Refused, bad username or password'
-                elif rc == mqtt_client.CONNACK_REFUSED_NOT_AUTHORIZED:
-                    reason = 'Connection Refused, not authorized'
-                LOGGER.error(f'{msg} {reason}')
-
-        self.conn.on_connect = on_connect
-
         self.conn.connect(self.broker_url.hostname, self._port)
-        LOGGER.debug('Connected to broker')
+        LOGGER.debug('Connection initiated')
 
     def pub(self, topic: str, message: str, qos: int = 1) -> bool:
         """
@@ -102,16 +80,26 @@ class MQTTPubSubClient(BasePubSubClient):
         :returns: `bool` of publish result
         """
 
-        self.conn.loop_start()
-        if self.is_connected:
+        def on_connect(client, userdata, flags, rc, properties=None):
+            if rc == 0:
+                LOGGER.debug('connected to broker')
+            else:
+                msg = 'Failed to connect to MQTT-broker:'
+                LOGGER.error(f'{msg} {mqtt_client.connack_string(rc)}')
+        self.conn.on_connect = on_connect   
+        # allow 10 attempts to connect
+        attempts = 0
+        while not self.conn.is_connected() and attempts < 10:
+            self.conn.loop()
+            time.sleep(1.)
+            attempts += 1
+        # publish message, if connected
+        if self.conn.is_connected():
             self.conn.publish(topic, message, qos)
-        self.conn.loop_stop()
-
-        if not self.is_connected:
-            LOGGER.error(f"Cannot publish to {self.broker['url']}, not connected") # noqa
-            return False
-        else:
             return True
+        else:
+            LOGGER.error(f'{self.client_id} failed to publish message')
+            return False
 
     def sub(self, topic: str) -> None:
         """
@@ -123,10 +111,14 @@ class MQTTPubSubClient(BasePubSubClient):
         """
 
         def on_connect(client, userdata, flags, rc):
-            LOGGER.debug(f'Connected to broker {self.broker}')
-            LOGGER.debug(f'Subscribing to topic {topic} ')
-            client.subscribe(topic, qos=1)
-            LOGGER.debug(f'Subscribed to topic {topic}')
+            if rc == 0:
+                LOGGER.debug(f'Connected to broker {self.broker}')
+                LOGGER.debug(f'Subscribing to topic {topic} ')
+                client.subscribe(topic, qos=1)
+                LOGGER.debug(f'Subscribed to topic {topic}')
+            else:
+                msg = 'Failed to connect to MQTT-broker:'
+                LOGGER.error(f'{msg} {mqtt_client.connack_string(rc)}')
 
         def on_disconnect(client, userdata, rc):
             LOGGER.debug(f'Disconnected from {self.broker}')
